remove domain and data models with associated mappers

+++ shopping_list/src/main/java/com/zizohanto/android/tobuy/shopping_list/presentation/mappers/ProductModelMapper.kt
@@ -1,6 +1,6 @@
 package com.zizohanto.android.tobuy.shopping_list.presentation.mappers
 
-import com.zizohanto.android.tobuy.domain.models.Product
+import com.zizohanto.android.tobuy.domain.sq.Product
 import com.zizohanto.android.tobuy.presentation.mapper.ModelMapper
 import com.zizohanto.android.tobuy.shopping_list.presentation.models.ProductsViewItem.ProductModel
 import javax.inject.Inject
@@ -14,7 +14,7 @@ class ProductModelMapper @Inject constructor() : ModelMapper<ProductModel, Produ
                 shoppingListId,
                 name,
                 price,
-                position
+                position.toInt()
             )
         }
     }
@@ -26,7 +26,7 @@ class ProductModelMapper @Inject constructor() : ModelMapper<ProductModel, Produ
                 shoppingListId,
                 name,
                 price,
-                position
+                position.toLong()
             )
         }
     }
